requirements.txt
python=3.6.8
h5py==2.10.0
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
Markdown==3.1.1
matplotlib==3.1.2
numpy==1.17.4
Pillow==6.2.1
scikit-image==0.16.2
scipy==1.3.3
tensorboard==2.0.1
torch==1.3.1
torchvision==0.4.2
tqdm==4.40.0
